Horizontal distribution button tool works!
Last prompt: You're getting closer.  There are 5 orange buttons on the bottom of the scene  view in this screenshot.   The three in the middle are not supposed to be overlapping.  All 5 buttons are supposed to be spaced evenly by the horizontal distributor tool.

Response: Ah, I see the issue now - we're dealing with UI buttons that need to be distributed evenly along the bottom of the screen. Let's modify the script to work specifically with RectTransform components and use anchoredPosition:...

Result:
The selected buttons are positioned accurately. Hopefully this will work with other items now.